+++ CHANGELOG.md
@@ -5,18 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2020-05-17
+
+### Added
+
+- Added S3 Engine
+
 ## [1.1.0] - 2019-12-15
 
 ### Added
 
-- Memory Engine
-- Zip Engine
+- Added Memory Engine
+- Added Zip Engine
 
 ## [1.0.1] - 2019-10-31
 
 ### Changed
 
-- Storage::list file result can be accessed as both array and object (FileObject)
+- Changed Storage::list file result can be accessed as both array and object (FileObject)
 
 ### Fixed